* origin/pr/22945:
  Generalizing to avoid warnings.
  Removed & to compile
  Converted to int in order to prevent unsigned underflow.
  Ran the files through astyles.
  Fix (some of) Travis' complaints.
  Attempting to not get Travis or Gorgon errors.
  Change unsigned long to size_t
  Made a lot of things unsigned for type safety.

+++ src/mod_manager.h
@@ -170,11 +170,11 @@ class mod_ui
 
         void try_add( const std::string &mod_to_add,
                       std::vector<std::string> &active_list );
-        void try_rem( int selection, std::vector<std::string> &active_list );
-        void try_shift( char direction, int &selection, std::vector<std::string> &active_list );
+        void try_rem( size_t selection, std::vector<std::string> &active_list );
+        void try_shift( char direction, size_t &selection, std::vector<std::string> &active_list );
 
-        bool can_shift_up( int selection, std::vector<std::string> active_list );
-        bool can_shift_down( int selection, std::vector<std::string> active_list );
+        bool can_shift_up( long selection, std::vector<std::string> active_list );
+        bool can_shift_down( long selection, std::vector<std::string> active_list );
 
     private:
         void set_usable_mods();
